renamed useFetch file to fetchJson
